Always try to initialize when the DOMContentLoaded event occurs.
In some old versions of Internet Explorer, 'document.readyState' may
be set to 'interactive' even while the document is still loading.[1]
This can even be the case before the <body> element has been loaded.

As a result, initialiseCKEditor() may be invoked too early, before the
textarea elements have been loaded.

To fix this:

- If document.body is undefined (i.e., ckeditor-init.js is loaded from
  <head>), ignore document.readyState and don't do any initialization
  until the DOMContentLoaded event occurs.

- If document.body exists and document.readyState is not 'loading',
  invoke initialiseCKEditor() and initialiseCKEditorInInlinedForms()
  immediately, but also set a listener to invoke initialiseCKEditor
  again, when and if the DOMContentLoaded event occurs.

In standards-compliant browsers, this shouldn't have any effect,
because DOMContentLoaded shouldn't be triggered again.  But in any
case, initialiseCKEditor() should be idempotent, so calling it twice
is harmless.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Document/readyState#Browser_compatibility